Added Login and logout user session management. Added logout, change password, tech support actions within app.

+++ src/org/nrjd/bv/app/util/CommonUtils.java
@@ -31,7 +31,7 @@ public class CommonUtils {
         }
     }
 
-    public static void sleep(int milliSecs) {
+    public static void sleep(long milliSecs) {
         sleep(null, milliSecs);
     }
